Make base box
Script was added to be able to use this box as a base box for the testing.

Had to change to bento, because the ubuntu/xenial64 box doesn't use the vagrant specified credentials. This means that any derived box will get authentication errors (it doesn't know to use the ubuntu user instead of the vagrant user).

Removed the port 5432 forward as it was decided to have users go through an SSH tunnel.